requirements.txt
fastapi
uvicorn
requests
PyGithub
